+++ EngineeringAcoustics.Curve/ComplexCurvePassthruMutator.cs
@@ -10,7 +10,7 @@ namespace EngineeringAcoustics.Curve
 	{
 		public virtual bool IsConstant => true;
 		public virtual string Name => "None";
-		public virtual float Mutate(float position, float input) => input;
+		public virtual double Mutate(double position, double input) => input;
 		public override string ToString() => Name;
 	}
 }
